requirements.txt
aiohttp>=3.8.6
botbuilder-core>=4.14.1
openai==0.28.1
python-dotenv>=1.0.1